Add files via upload
The bot allows you to buy any newly added token from PancakwSwap (all BSC network) and Uniswap in amazing speeds ( up to zero seconds from the time of the liquidity transaction show up in the mempool). It is called “frontrunning” which can make you a big amount of profit. The bot works by: Connecting to the blockchain network, scanning for all mempool / pending transactions, finding and matching the liquidity add transaction of the token you are trying to snipe and then placing the buy order in the same block